requirements.txt
requests>=2.22
beautifulsoup4==4.8.0
Pillow==6.2.0
python>=3.7
